+++ examplepatches/humzakh/HK-Morphe-Patches/extensions/syncforreddit/src/main/java/app/morphe/extension/syncforreddit/RandomSubredditExtension.java
@@ -7,7 +7,7 @@ import org.json.JSONObject;
  *
  * @noinspection unused
  */
-public class FixRandomSubredditPatch {
+public class RandomSubredditExtension {
     public static String parseRandomResponse(byte[] data) {
         try {
             String responseBody = new String(data, "UTF-8");