+++ examplepatches/humzakh/HK-Morphe-Patches/extensions/syncforreddit/stubs/src/main/java/nc/a.java
@@ -1,0 +1,11 @@
+package nc;
+
+public class a {
+    public static a b() {
+        return null;
+    }
+
+    public static a c() {
+        return null;
+    }
+}