+++ examplepatches/humzakh/HK-Morphe-Patches/patches/src/main/kotlin/app/template/patches/reddit/customclients/sync/syncforreddit/fix/ultra/Fingerprints.kt
@@ -1,7 +1,5 @@
 package app.template.patches.reddit.customclients.sync.syncforreddit.fix.ultra
 
-import app.template.patches.reddit.customclients.sync.syncforreddit.SyncForRedditCompatible
-
 import app.morphe.patcher.Fingerprint
 import com.android.tools.smali.dexlib2.AccessFlags